Created hello.java
This specific program prints "Hello!", on the output screen.
The package I have created was for me and so in order to print this program and get the correct output without any syntax error, The user has to either create a same named package "Hellojava", on their programming platform or simply remove the package statement from my program. Then compile it and they will get the correct output, "Hello!".
Thank you!